OK-679: Siirretty isLaskentaUsedForValinnanvaihe valintaperusteet-utils-tiedostoon ja vaihdettu nimi kuvaavammaksi
